MQTT.jl

MQTT.Client — Type

Client(on_message, on_disconnect, [ping_timeout=60])

The client struct is used to store state for a MQTT session

Arguments

on_message::Function: function to be called upon receiving a publish message

on_disconnect::Function: function to be called when disconnected

ping_timeout::UInt64=60: seconds the client waits for the PINGRESP after sending a PINGREQ before he disconnects

MQTT.ConnectOpts — Type

ConnectOpts(host, [port=1883]; <keyword arguments>)
ConnectOpts(get_io; <keyword arguments>)

Create a ConnectOpts using the given hostname and port of the broker.

Arguments

host::AbstractString: host of the broker

port::Integer=1883: port of the broker

get_io::Function: function to create a new IO object (e.g. a TCPSocket)

Keywords

username::Union{Nothing, String} = nothing: username as a UTF-8 string

password::Union{Nothing, Vector{UInt8}} = nothing: password as a vector of bytes

client_id::String = "": client id as a UTF-8 string. Zero-length string as an id is valid per MQTT 3.1.1 spec

clean_session::Bool = true: discard any stored session and create a new one

keep_alive::UInt16 = 0x0000: set maximum keep alive interval

will::Union{Nothing, Message} = nothing: a Message containing a will

Examples

opts = ConnectOpts(
    "localhost";
    username="myuser",
    password=Vector{UInt8}("secretpassword"),
    client_id="my_mqtt_client"
)

MQTT.Message — Type

Message(topic, payload; [dup=false], [qos=AT_MOST_ONCE], [retain=false])

A container for a message with all required metadata.

Can be used as a last will message.

Arguments

topic::String

payload::Vector{UInt8}

Keywords

dup::Bool: message is a duplicate

qos::QoS: QoS level with which the message should be send

retain::Bool: broker should retain the message

Base.get — Method

get(future)

Connects the Client instance using the given options.

MQTT.disconnect — Function

disconnect(client, [reason=nothing])

Disconnects the Client instance gracefully, shuts down the background tasks and stores session state.

Arguments

client::Client: client to disconnect

reason::Union{Exception,Nothing}=nothing: reason for the disconnect This is forwarded to the on_disconnect callback

MQTT.publish — Method

publish(client, topic, payload; <keyword arguments>)

Publishes a message to the broker connected to the Client instance provided as a parameter.

Arguments

client::Client: The client to send the message over.

topic::String: The topic to publish on. Normal rules for publish topics apply so "/" are allowed but no wildcards.

payload::Union{String, Vector{UInt8}}: payload to deliver. This should be a Vector{UInt8}. String is automatically converted. A zero length payload is allowed.

Keywords

async::Bool=false: If false the call blocks until the operation is complete.

qos::QoS=AT_MOST_ONCE: The MQTT quality of service to use for the message. This has to be a QoS constant (ATMOSTONCE, ATLEASTONCE, EXACTLY_ONCE).

retain::Bool=false: Whether or not the message should be retained by the broker. This means the broker sends it to all clients who subscribe to this topic

MQTT.subscribe — Method

subscribe(client, topics...; [async=false])
subscribe(client, topics...; [qos=AT_MOST_ONCE], [async=false])

Subscribes the Client instance, provided as a parameter, to the specified topics.

Arguments

client::Client: the client to subscribe to the topics

topics...: Topics to connect to. This can either be a list of Tuple{String, QoS} where the first element is a topic name and the second a QoS level, or a list of String topic names (AT_MOST_ONCE is implied unless otherwise specified with the qos::QoS keyword)

Keywords

async::Bool=false: If false the call blocks until the operation is complete

qos::QoS=AT_MOST_ONCE: QoS level to subscribe with

MQTT.unsubscribe — Method

unsubscribe(client, topics...; [async=false])

This method unsubscribes the Client instance from the specified topics.

Arguments

client::Client: the client to unsubscribe from the topics

topics::String...: list of topics to unsubscribe from

Keywords

async::Bool=false: If false the call blocks until the operation is complete.

Sockets.connect — Method

connect(client, opts; [async=false])

Connects to a broker using the specified options.

Arguments

client::Client: client needs to be instantiated before connecting, but can be reconnected using connect

opts::ConnectOpts: connection options to be used

Keywords

async::Bool: if true a Future is returned, therwise the function blocks until the client is connected or the operation failed.

See also get.