MQTT.AbstractConnection

MQTT.QOS

MQTT.AbstractConnection — Type

MQTTConnection

An abstract type representing a connection to an MQTT broker.

MQTT (Message Queuing Telemetry Transport) is a lightweight, publish-subscribe, machine-to-machine network protocol for message queue/message queuing service. It is designed for connections with remote locations that have devices with resource constraints or limited network bandwidth, such as in the Internet of Things (IoT). It must run over a transport protocol that provides ordered, lossless, bi-directional connections—typically, TCP/IP.

An MQTT client establishes a connection with the MQTT broker. Once connected, the client can either publish messages, subscribe to specific messages, or do both. When the MQTT broker receives a message, it forwards it to subscribers who are interested.

MQTT.QOS — Type

QOS

An enum representing the different Quality of Service (QoS) levels in MQTT.

Values

AT_MOST_ONCE: QoS level 0, at most once delivery. The message is delivered at most once, or it may not be delivered at all. This is also known as "fire and forget".

AT_LEAST_ONCE: QoS level 1, at least once delivery. The message is guaranteed to be delivered at least once, but it may be delivered multiple times.

EXACTLY_ONCE: QoS level 2, exactly once delivery. The message is guaranteed to be delivered exactly once.

Missing docstring.

Missing docstring for connect. Check Documenter's build log for details.

Missing docstring.

Missing docstring for connect!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for subscribe. Check Documenter's build log for details.

Missing docstring.

Missing docstring for subscribe!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for publish. Check Documenter's build log for details.

Missing docstring.

Missing docstring for publish!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for unsubscribe. Check Documenter's build log for details.

Missing docstring.

Missing docstring for unsubscribe!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for disconnect. Check Documenter's build log for details.

Missing docstring.

Missing docstring for disconnect!. Check Documenter's build log for details.